GH-32 customer orders a Product, not a cookie

+++ src/main/java/fr/unice/polytech/si4/otake/cookiefactory/RegisteredCustomer.java
@@ -26,7 +26,7 @@ public class RegisteredCustomer {
 	 * @param price
 	 * @return the new price or the old price if not eligible
 	 */
-	public float addDiscountIfEligible(float price) {
+	public double addDiscountIfEligible(double price) {
 		float d = (float) 0.;
 		if (unitsOfCookiesBeforeDiscount <= 0) {
 			d = this.discount;
